[Fix Warning] sign-compare Warning: QTAtom_stsd.cpp:118:3: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]

+++ QTFileLib/QTAtom_stsd.cpp
@@ -115,7 +115,7 @@ Bool16 QTAtom_stsd::Initialize(void)
         //
         // Allocate our description tables.
         UInt64 tableSize = fTOCEntry.AtomDataLength - 8;
-		Assert(tableSize < kSInt32_Max);
+        Assert(tableSize < (UInt64)kSInt32_Max);
         fSampleDescriptionTable = NEW char[ (SInt32) tableSize];
         if( fSampleDescriptionTable == NULL )
             return false;
